+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Infra:
 
   - running UI tests in CircleCI
   - getting extension ID dynamically to run UI tests
-  - [todo] changing build strategy for extension, generate manifest dynamically for tests
+  - changing build strategy for extension, generate manifest dynamically for UI tests
   - add test result in circleci with jest junit
 
 Tests: